+++ HISTORY.rst
@@ -32,6 +32,7 @@ End-User Summary
 - Added flags `segregates`, `doesnt_segregate` and `no_disease_association` to file export (#502).
 - Adjusting path to new varfish-annotator db download (#546).
 - Adding feature to enable and configure link-out to HGMD (#576).
+- Small variant filtration results now allow to easily look up second hits in the same gene (#573).
 
 Full Change List
 ================
@@ -63,6 +64,7 @@ Full Change List
 - Adjusting path to new varfish-annotator db download (#546).
 - Fixing issue with sync-from-remote when no remote is defined (#570).
 - Adding feature to enable and configure link-out to HGMD (#576).
+- Small variant filtration results now allow to easily look up second hits in the same gene (#573).
 
 ------
 v1.2.0